Move logging related tasks to its own package.
