+++ main.go
@@ -1,18 +1,11 @@
 package main
 
 import (
-	"crypto/tls"
-	"fmt"
-	log "github.com/Sirupsen/logrus"
-	"github.com/Sirupsen/logrus/hooks/syslog"
 	"github.com/SpringerPE/firehose-to-syslog/firehose"
+	"github.com/SpringerPE/firehose-to-syslog/logging"
 	"github.com/SpringerPE/firehose-to-syslog/token"
-	"github.com/cloudfoundry/noaa"
 	"github.com/cloudfoundry/noaa/events"
 	"gopkg.in/alecthomas/kingpin.v1"
-	"io/ioutil"
-	"log/syslog"
-	"os"
 )
 
 var (
@@ -26,179 +19,27 @@ var (
 	skipSSLValidation = kingpin.Flag("skip-ssl-validation", "Please don't").Bool()
 )
 
-
 func RouteEvents(in chan *events.Envelope) {
 	for msg := range in {
 		switch msg.GetEventType() {
 		case events.Envelope_Heartbeat:
-			Heartbeats(msg)
+			logging.Heartbeats(msg)
 		case events.Envelope_HttpStart:
-			HttpStarts(msg)
+			logging.HttpStarts(msg)
 		case events.Envelope_HttpStop:
-			HttpStops(msg)
+			logging.HttpStops(msg)
 		case events.Envelope_HttpStartStop:
-			HttpStartStops(msg)
+			logging.HttpStartStops(msg)
 		case events.Envelope_LogMessage:
-			LogMessages(msg)
+			logging.LogMessages(msg)
 		case events.Envelope_ValueMetric:
-			ValueMetrics(msg)
+			logging.ValueMetrics(msg)
 		case events.Envelope_CounterEvent:
-			CounterEvents(msg)
+			logging.CounterEvents(msg)
 		case events.Envelope_Error:
-			ErrorEvents(msg)
+			logging.ErrorEvents(msg)
 		case events.Envelope_ContainerMetric:
-			ContainerMetrics(msg)
-		}
-	}
-}
-
-func Heartbeats(msg *events.Envelope) {
-	heartbeat := msg.GetHeartbeat()
-
-	log.WithFields(log.Fields{
-		"ctl_msg_id":     heartbeat.GetControlMessageIdentifier(),
-		"error_count":    heartbeat.GetErrorCount(),
-		"event_type":     msg.GetEventType(),
-		"origin":         msg.GetOrigin(),
-		"received_count": heartbeat.GetReceivedCount(),
-		"sent_count":     heartbeat.GetSentCount(),
-	}).Info("")
-}
-
-func HttpStarts(msg *events.Envelope) {
-	httpStart := msg.GetHttpStart()
-
-	log.WithFields(log.Fields{
-		"event_type":        msg.GetEventType(),
-		"origin":            msg.GetOrigin(),
-		"cf_app_id":         httpStart.GetApplicationId(),
-		"instance_id":       httpStart.GetInstanceId(),
-		"instance_index":    httpStart.GetInstanceIndex(),
-		"method":            httpStart.GetMethod(),
-		"parent_request_id": httpStart.GetParentRequestId(),
-		"peer_type":         httpStart.GetPeerType(),
-		"request_id":        httpStart.GetRequestId(),
-		"remote_addr":       httpStart.GetRemoteAddress(),
-		"timestamp":         httpStart.GetTimestamp(),
-		"uri":               httpStart.GetUri(),
-		"user_agent":        httpStart.GetUserAgent(),
-	}).Info("")
-}
-
-func HttpStops(msg *events.Envelope) {
-	httpStop := msg.GetHttpStop()
-
-	log.WithFields(log.Fields{
-		"event_type":     msg.GetEventType(),
-		"origin":         msg.GetOrigin(),
-		"cf_app_id":      httpStop.GetApplicationId(),
-		"content_length": httpStop.GetContentLength(),
-		"peer_type":      httpStop.GetPeerType(),
-		"request_id":     httpStop.GetRequestId(),
-		"status_code":    httpStop.GetStatusCode(),
-		"timestamp":      httpStop.GetTimestamp(),
-		"uri":            httpStop.GetUri(),
-	}).Info("")
-}
-
-func HttpStartStops(msg *events.Envelope) {
-	httpStartStop := msg.GetHttpStartStop()
-
-	log.WithFields(log.Fields{
-		"event_type":        msg.GetEventType(),
-		"origin":            msg.GetOrigin(),
-		"cf_app_id":         httpStartStop.GetApplicationId(),
-		"content_length":    httpStartStop.GetContentLength(),
-		"instance_id":       httpStartStop.GetInstanceId(),
-		"instance_index":    httpStartStop.GetInstanceIndex(),
-		"method":            httpStartStop.GetMethod(),
-		"parent_request_id": httpStartStop.GetParentRequestId(),
-		"peer_type":         httpStartStop.GetPeerType(),
-		"remote_addr":       httpStartStop.GetRemoteAddress(),
-		"request_id":        httpStartStop.GetRequestId(),
-		"start_timestamp":   httpStartStop.GetStartTimestamp(),
-		"status_code":       httpStartStop.GetStatusCode(),
-		"stop_timestamp":    httpStartStop.GetStopTimestamp(),
-		"uri":               httpStartStop.GetUri(),
-		"user_agent":        httpStartStop.GetUserAgent(),
-	}).Info("")
-}
-
-func LogMessages(msg *events.Envelope) {
-	logMessage := msg.GetLogMessage()
-
-	log.WithFields(log.Fields{
-		"event_type":      msg.GetEventType(),
-		"origin":          msg.GetOrigin(),
-		"cf_app_id":       logMessage.GetAppId(),
-		"timestamp":       logMessage.GetTimestamp(),
-		"source_type":     logMessage.GetSourceType(),
-		"message_type":    logMessage.GetMessageType().String(),
-		"source_instance": logMessage.GetSourceInstance(),
-	}).Info(string(logMessage.GetMessage()))
-}
-
-func ValueMetrics(msg *events.Envelope) {
-	valMetric := msg.GetValueMetric()
-
-	log.WithFields(log.Fields{
-		"event_type": msg.GetEventType(),
-		"origin":     msg.GetOrigin(),
-		"name":       valMetric.GetName(),
-		"unit":       valMetric.GetUnit(),
-		"value":      valMetric.GetValue(),
-	}).Info("")
-}
-
-func CounterEvents(msg *events.Envelope) {
-	counterEvent := msg.GetCounterEvent()
-
-	log.WithFields(log.Fields{
-		"event_type": msg.GetEventType(),
-		"origin":     msg.GetOrigin(),
-		"name":       counterEvent.GetName(),
-		"delta":      counterEvent.GetDelta(),
-		"total":      counterEvent.GetTotal(),
-	}).Info("")
-}
-
-func ErrorEvents(msg *events.Envelope) {
-	errorEvent := msg.GetError()
-
-	log.WithFields(log.Fields{
-		"event_type": msg.GetEventType(),
-		"origin":     msg.GetOrigin(),
-		"code":       errorEvent.GetCode(),
-		"delta":      errorEvent.GetSource(),
-	}).Info(errorEvent.GetMessage())
-}
-
-func ContainerMetrics(msg *events.Envelope) {
-	containerMetric := msg.GetContainerMetric()
-
-	log.WithFields(log.Fields{
-		"event_type":     msg.GetEventType(),
-		"origin":         msg.GetOrigin(),
-		"cf_app_id":      containerMetric.GetApplicationId(),
-		"cpu_percentage": containerMetric.GetCpuPercentage(),
-		"disk_bytes":     containerMetric.GetDiskBytes(),
-		"instance_index": containerMetric.GetInstanceIndex(),
-		"memory_bytes":   containerMetric.GetMemoryBytes(),
-	}).Info("")
-}
-
-func setupLogging(syslogServer string, debug bool) {
-	log.SetFormatter(&log.JSONFormatter{})
-	log.SetOutput(os.Stdout)
-	if !debug {
-		log.SetOutput(ioutil.Discard)
-	}
-	if syslogServer != "" {
-		hook, err := logrus_syslog.NewSyslogHook("tcp", syslogServer, syslog.LOG_INFO, "doppler")
-		if err != nil {
-			log.Error("Unable to connect to syslog server.")
-		} else {
-			log.AddHook(hook)
+			logging.ContainerMetrics(msg)
 		}
 	}
 }
@@ -207,8 +48,7 @@ func main() {
 	kingpin.Version("0.0.2 - ba541ca")
 	kingpin.Parse()
 
-	setupLogging(*syslogServer, *debug)
-
+	logging.SetupLogging(*syslogServer, *debug)
 	token := token.GetToken(*uaaEndpoint, *firehoseUser, *firehosePassword, *skipSSLValidation)
 	firehose := firehose.CreateFirehoseChan(*dopplerEndpoint, token, *subscriptionId, *skipSSLValidation)
 	RouteEvents(firehose)
